pyproject.toml
[tool.poetry]
name = "voiceai"
version = "0.3.0"
description = "Real-time AI Voice Assistant with Character Persona"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.8.0"
python-dotenv = "^1.0.0"
rich = "^13.7.0"
edge-tts = "^6.1.0"
pydub = "^0.25.0"
aiofiles = "^23.0.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
websockets = "^14.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"

[tool.poetry.scripts]
voiceai = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
